fastapi/requirements.txt
fastapi
uvicorn[standard]
pandas
gunicorn
pydantic
scikit-surprise